Moving to EmailHistory file
